backend/src/requirements.txt
boto3==1.35.69
botocore==1.35.69
PyJWT==2.9.0
python-dateutil==2.9.0